move DeleteProfile file into a folder
