+++ client/src/components/profile/profile_header/ProfileHeader.jsx
@@ -2,7 +2,7 @@ import PropTypes from "prop-types";
 // import Pen from "../../../assets/images/penblue.png";
 import LogoCicorne from "../../../assets/logo/cicorne.png";
 import styles from "./ProfileHeader.module.css";
-import DeleteProfile from "../DeleteProfile";
+import DeleteProfile from "./delete_profile/DeleteProfile";
 
 function ProfileHeader({
   username,
